Remove SloppyVertex_Vertex dependence on module-level stencils
They're fast enough to construct one time.  But the upcoming tweaks to
the volume-averaged non-sloppy measurement stores a copy of the path P
for every starting point to reduce data movement.  And that's a
different shape than the one-measurement stencil, so to avoid collisions
better to make the sloppy one slow.