Fix the symmetrization
The symmetrization was lifted from tdg where, for reasons of ergodicity,
we only do N=odd.  But here we do N=even too.  There was a bug for
N=even.

The bug was that the set of coordinates for an even-sized lattice is not
closed under a naive application of the symmetry group operations.

However, if we mod the coordinates back into the lattice all is fixed.