NotImplemented --> NotImplementedError (#75)
These are not the same, as I found by trying to handle exceptions and just raising more.

NotImplemented is a sentinel singleton designed to be returned in
unimplemented comparators to indicate the opposite comparison should
be tried.  Importantly it is not callable!  It doesn't raise an exception!

NotImplementedError is the exception.