Simple but costly change to PlaquetteUpdate
The reason it is simple is that it computes δv and then just makes the
obvious m --> m - δv/W substitution in ds.  However, computing δv is a
volume cost.  This is quite wasteful because as v changes δv changes, so
the volume cost is inside a local update.

This will be fixed in an upcoming commit wherein we get explicit forms
for δv on each link and get an O(1) calculation of dS rather than O(V).