metropolis.py --> villain.py
The NeighborhoodUpdate was previously imported at the generator module
level, which makes no sense (unless we were going to import EVERY
generator at the module level).  So now to get the neighborhoodupate
you've got to use `supervillain.generator.villain.NeighborhoodUpdate`
instead of just `supervillain.generator.NeighborhoodUpdate`.

This change was made in many examples and the documentation.