Observables have autocorrelation() classmethod
Currently all observables are NOT included in the autocorrelation_time
of ensembles, but we will update observables shortly.

It has to be a method because deciding requires knowing the ensemble (to
test W, for example).  Making it a classmethod gives it access to
potentially greater decision-making power.  It can't be a standard
method because Observables aren't instantiated.