Feature/better autocorrelation (#93)
Observables know whether they should be used to compute and ensemble's autocorrelation_time().
There are a variety of reasons to not use every observable.  For example, it's trickier to do it for spacetime-dependent correlation functions---not impossible but currently out of scope.  Or, when W=1, some vortex observables are trivially independent of configuration and look like they have infinite autocorrelation.  But it'd be wrong to run forever---it's just that the observable is dumb when W=1.

To make it simple to compute an ensemble's autocorrelation times we made improvements to the Observable class and particular observables.

* Register observables in a single dict
* Observables have autocorrelation() classmethod
    It has to be a method because deciding requires knowing the ensemble (to
    test W, for example).  Making it a classmethod gives it access to
    potentially greater decision-making power.  It can't be a standard
    method because Observables aren't instantiated; Constrained needs super()
* Observables might also be Scalar or Constrained.
    It's easy to use Scalars for autocorrelation time estimates but Constrained
    observables are completely autocorrelated when W=1.  So don't use them when W=1.
* Update extant observables with Constrained, Scalar
* Ensemble.measure is great shorthand
* Examples use .measure and .autocorrelation_time
* Point analysis docs at ensemble.autocorrelation_time