Inline worm measurements are right!
We can see that they're right by measuring eg. Vortex_Vortex in a
non-trivial case.  That's what example/why-measure-lineline.py does.

To really honestly match the other formulation, we needed to adjust the
worm tallying.  I think it makes sense now: if we put the worm down and
decide to immediately take it away, we should think of that as having a
worm of length 0, so the tally should be 0.