Classic worldline worm
We really struggled to understand why the Geometric worm did not produce
exactly the right correlator (which we know by computing in the dual
formulation).  So I implemented the classic worm.  Slower than the
Geometric one, but it gives the right correlator, so perhaps something
is wrong with our Geometric worm.