Remove the Geometric worm in favor of the Classic
After some detailed computations in both duality frames we concluded
that something must be wrong with our implementation of the Geometric
worm which caused observables to not match across the duality.

This caused me to go back and revisit the explanation of the worm, which
is now included in the documentation, and which led to a very
straightforward adaptation of the original Prokof'ev and Svistunov
algorithm to both the Villain and Worldline framing.