Feature/critical moments (#136)
ALGORITHMS
----------
* Document inline Spin_Spin and Vortex_Vortex worm observables.

 - Normalize by the constraint-satisfying worms on
   {Spin_Spin,Vortex_Vortex}{Susceptibility,CriticalMoment}
 - Explain why the normalization has to happen on expectation values.
 - Change all the reduction observables into DerivedQuantities.
* Classic worldline worm

We really struggled to understand why the Geometric worm did not produce
exactly the right correlator (which we know by computing in the dual
formulation).  So I implemented the classic worm.  Slower than the
Geometric one, but it gives the right correlator, so perhaps something
is wrong with our Geometric worm.

* Worldline worm comments + documentation
* Explain worm clearly in docs
* Classic worms implement the obvious algorithm.
* Remove the Geometric worm in favor of the Classic

After some detailed computations in both duality frames we concluded
that something must be wrong with our implementation of the Geometric
worm which caused observables to not match across the duality.

This caused me to go back and revisit the explanation of the worm, which
is now included in the documentation, and which led to a very
straightforward adaptation of the original Prokof'ev and Svistunov
algorithm to both the Villain and Worldline framing.


GENERATORS
----------

* Generator class + documentation
* Configurations get inline observables during generation

 - Add |= (__ior__) for configurations
 - Add in (__contains__) for configurations
 - Add inline_observables into Ensemble.generate
 - .copy() ensures that inline measurements are copied over; tests are updated to use the copy method
* Extant generators are Generators
* Observables check for inline measurements before computing


ANALYSIS
--------

* Blocking

Sometimes the vortex vortex correlator has HUGE spikes (kappa, W, and
lattice dependent), which makes sense: to get a polynomial falloff from
an observable which is naively an exponentiated difference in action,
large tails are needed to balance typically-exponentially-small
measurements.

The issue is that the spikes can be very rare, and taking only .every
nth measurement can remove them completely, leading to bias.  @j-ostmeyer
suggested blocking instead of decimating, since averaging over a block
automatically includes the spikes---you don't have to 'get lucky' to
include them in a subsequent bootstrap.

* Point to Blocking.plot_history

I tried making Blocking inherit from Ensemble but a number of
problematic things happened---I wasn't capturing the observable calls
correctly, so that calling blocking.observable was measuring instead of
averaging over the ensemble's measurements.

Assigning Blocking.plot_history = Ensemble.plot_history created either
circular dependency problems or redirection problems.

OBSERVABLES
-----------

* Overlay Vortex_Vortex Villain correlator
* Relocate the CriticalScalingDimension

Since it will be used not just by the SpinSusceptibilityScaled it makes
sense to attach it to a more primitive observable.

* SpinCriticalMoment
* Relocate the Vortex CriticalScalingDimension
* VortexCriticalMoment
* Critical moments are .real
* `{Spin,Vortex}Susceptibility{,Scaled}` are DQs to allow for normalization of inline measurements.

TESTS + EXAMPLES
----------------
* Ignore inline measurements when comparing to reference implementations.

They may not match!  That's not the point of the reference
implementations anyway!
* Tests should be able to measure without relying on inline observables
* Remove SpinSusceptibility from examples; it is now DQ
* example/why-measure-inline.py allows us to contrast the inline measurements, the taxicab measurements, and the measurements on the other side of the duality.

FIXES + TWEAKS
--------------

* Fix off-by-1 error for even lattices
* Edge case where only non-autocorrelation observables are measured

For example, if Vortex_Vortex is measured manually but no scalars are,
the auto dict was previously empty and so the max was invalid.

* Action shortcuts
* Use ensemble's len
* Clean up left-over generators
* Plotting multiplier to make observable comparisons easier
* L.R_squared is the square distance to the origin, useful when computing the critical moments.