Harmonize worm tallying
This isn't a bugfix exactly, but it does affect the matching between
self-dual W=∞ computations.  The issue is essentially that by tallying
at different times we slightly change the histograms of the inline
observables.  In particular, different choices slightly change the tally
at 0 displacement, resulting in differing normalizations.

This harmonizes those choices so that the self-dual correlator match as
close as we care to measure.