Ensemble tau defaults to 2
Assuming the thermalization production has done its job, the
autocorrelation time will be short.  But if physically some observables
are very very small (like TorusWrappings) you might not get enough
fluctuation in an ensemble.  If that observable was what was slow during
thermalization, you might just get unlucky in terms of its evolution
in strides of τ: the odds of seeing a change are like 1/τ.